+++ source/common/router/router.h
@@ -195,6 +195,7 @@ public:
   const Network::Connection* downstreamConnection() const override {
     return callbacks_->connection();
   }
+  const Http::HeaderMap* downstreamHeaders() const override { return downstream_headers_; }
 
   /**
    * Set a computed cookie to be sent with the downstream headers.