+++ source/common/ssl/context_config_impl.h
@@ -49,6 +49,7 @@ public:
   const std::vector<std::string>& verifyCertificateSpkiList() const override {
     return verify_certificate_spki_list_;
   };
+  bool allowExpiredCertificate() const override { return allow_expired_certificate_; };
   unsigned minProtocolVersion() const override { return min_protocol_version_; };
   unsigned maxProtocolVersion() const override { return max_protocol_version_; };
 
@@ -78,6 +79,7 @@ private:
   const std::vector<std::string> verify_subject_alt_name_list_;
   const std::vector<std::string> verify_certificate_hash_list_;
   const std::vector<std::string> verify_certificate_spki_list_;
+  const bool allow_expired_certificate_;
   const unsigned min_protocol_version_;
   const unsigned max_protocol_version_;
 };